simplfy dukeds backend
Moves utility methods from .backends.dukeds.py to .oauth_utils.py
This removes a dependency where gcb_web_auth/oauth_utils.py depended upon gcb_web_auth/backends/dukeds.py.
